Add StatusReport related things, now you can get, add status reports

+++ src/main/java/hu/dd/varianddbackend/controller/PatientController.java
@@ -23,7 +23,7 @@ public class PatientController {
     }
 
     @GetMapping("/patient/{id}")
-    public Patient getPatientById(@PathVariable("id") Integer id)
+    public Patient getPatientById(@PathVariable("id") Long id)
     {
         return patientRepository.findById(id).orElseThrow(()-> {
             throw new ResponseStatusException(
